refactor: Badge domain repackage
